chore(share): move out reconstruction tests into a separate file with disabled race detector
